refactor: move test key manager to test utils folder.
