+++ tests/serverpod_test_server/lib/test_util/test_key_manager.dart
@@ -1,0 +1,18 @@
+import 'package:serverpod_test_client/serverpod_test_client.dart';
+
+class TestAuthKeyManager extends AuthenticationKeyManager {
+  String? _key;
+
+  @override
+  Future<String?> get() async => _key;
+
+  @override
+  Future<void> put(String key) async {
+    _key = key;
+  }
+
+  @override
+  Future<void> remove() async {
+    _key = null;
+  }
+}
