+++ tests/serverpod_test_server/test_e2e/websocket_test.dart
@@ -2,9 +2,8 @@ import 'package:serverpod_test_client/serverpod_test_client.dart';
 import 'package:serverpod_test_module_client/module.dart';
 import 'package:serverpod_test_client/src/custom_classes.dart';
 import 'package:serverpod_test_server/test_util/config.dart';
+import 'package:serverpod_test_server/test_util/test_key_manager.dart';
 import 'package:test/test.dart';
-
-import 'authentication_test.dart';
 
 void main() {
   var client = Client(
